TST: isolate rocket unit tests
